Split smells.py into Fowler sub-category modules
Replaces the monolithic 1857-line smells.py (25 rules) with five
focused modules organized along Fowler's own taxonomy:

- bloaters.py: LongFunction, LongParameterList, LargeClass,
  PrimitiveObsession, DataClumps
- oo_abusers.py: RepeatedSwitches, TemporaryField, RefusedBequest,
  AlternativeInterfaces
- change_preventers.py: DivergentChange, ShotgunSurgery,
  DuplicatedCode (with _Normalizer helper)
- dispensables.py: LazyElement, SpeculativeGenerality, Comments,
  DataClassSmell (owns the shared _BOILERPLATE_DUNDERS frozenset)
- couplers.py: FeatureEnvy, MessageChains, MiddleMan, InsiderTrading
- smells.py (remaining): MysteriousName, GlobalData, Loops, MutableData

Each module exports a *_RULES tuple aggregated in
packs/python/rules/__init__.py. No rule logic changes - pure
structural refactor.

Fixes #46
